fix user type detection for viewers
The @user_type tag gets parsed as "true" if no value is given,
which in turn meant that the user level check would become
"undefined < neededLevel", which is always false. Thus normal
viewers were allowed to do aaaanything!

Also rewrote the if condition to deny by default if a user type is
not known.